trading_bot/requirements.txt
numpy
pandas
matplotlib
scikit-learn
requests
ta-lib
torch
pymysql
python-dotenv